+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - added option to choose label design for delivery methods
 - added orders selection from multiple pages in Omniva Shipping page
 - labels are no longer stored in the plugin folder for a long time
+- added a ability to set custom shipping method name
 
 ## [1.9.2.1]
 ### Fixed